Project 5
First taste of Dynamic Programming coding. Some parts were hard to
conceptualize, so I wound up writing psuedocode for the later parts.

The whole point of this project was described as follows:
3. Consider the problem of neatly printing a paragraph on the screen (or
on a printer). The input text is a sequence
S of n words (represented as strings) of lengths `1; : : : ; `n
(measured in characters). The input bound M is the
maximum number of characters a line can hold. The key to neatly printing
a paragraph is to identify in the text
sequence the lines of the paragraph so that new-lines can be placed at
the end of each line. We can formalize
the notion of the “badness” of a line as the number of extra space
characters at the end of the line or 1 if the
bound M is exceeded. We can formalize the notion of the “badness” of a
paragraph as the badness of the worst
(i.e., maximum) line of the paragraph not including the last line. Thus
to identify the lines for a neat paragraph,
we seek to minimize the badness of the paragraph.
(a) If a given line contains words i through j, and we leave exactly one
space between words, the number
of extra space characters at the end of the line is M 􀀀 j + i 􀀀
Pj
k=i `k. Write a mathematical function
s(S;M; i; j) that computes the number of extra space characters at the
end of a line.
(b) (project) Write a procedure extraSpace that implements s.
(c) Use the function s to write a mathematical function bl(S;M; i; j)
that computes line badness.
(d) (project) Write a procedure badnessLine that implements bl.
(e) Write a recursive mathematical function mb(S;M) that computes the
minimum paragraph badness (using
slicing).
(f) Write a recursive mathematical function mb0(S;M; i) where mb0(S;M;
i) = mb(S[i :];M).
(g) (project) Write a recursive procedure minBad that implements the
function mb0. It should take three
parameters: S, M, and i a slicing index.
1
(h) (project) Write a procedure minBadDynamic that implements the
function mb0 using dynamic programming.
It should take only two parameters: S, and M.
(i) (project) Write a procedure minBadDynamicChoice that implements the
function mb0 using dynamic
programming. In addition to returning mb(S;M), it should also return the
choices made. Then write a
procedure printParagraph which takes two parameters: S and M that
displays the words in S on
the screen using the choices of minBadDynamicChoice. What is the
asymptotic running time of your
algorithm?